stuck on esp32 code for serial for now, had led working but not intergrated with class code. their is some type of issue with pin definitions maybe

+++ Microcontrollers/sunspotmc/include/buttons.h
@@ -8,7 +8,7 @@ class CapacitiveButton {
 private:   
     int pinSend;
     int pinReceive;
-    long threshold;
+    uint16_t threshold;
 
 public:
     CapacitiveButton(int sendPin,long thresholdValue);
